Добавляет store/reducer и store/action
