+++ src/store/action.js
@@ -1,0 +1,5 @@
+export const ActionType = {
+  SET_CITIES: `temp/city`,
+  SET_HOTELS: `temp/hotel`,
+  SET_ACTIVE_CITY: `SET_ACTIVE_CITY`,
+};
